s3: remove AbstractContentApi
